+++ src/main/java/org/jolokia/docker/maven/BuildMojo.java
@@ -7,6 +7,7 @@ import org.apache.maven.archiver.MavenArchiveConfiguration;
 import org.apache.maven.execution.MavenSession;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.shared.filtering.MavenFileFilter;
+import org.apache.maven.shared.filtering.MavenReaderFilter;
 import org.jolokia.docker.maven.access.DockerAccess;
 import org.jolokia.docker.maven.access.DockerAccessException;
 import org.jolokia.docker.maven.assembly.DockerAssemblyManager;
@@ -38,6 +39,9 @@ public class BuildMojo extends AbstractDockerMojo {
     /** @component */
     private MavenFileFilter mavenFileFilter;
 
+    /** @component */
+    private MavenReaderFilter mavenFilterReader;
+    
     /** @component */
     private DockerAssemblyManager dockerAssemblyManager;
 
@@ -74,7 +78,7 @@ public class BuildMojo extends AbstractDockerMojo {
 
         checkImageWithAutoPull(dockerAccess, fromImage, new ImageName(fromImage).getRegistry());
 
-        MojoParameters params =  new MojoParameters(session, project, archive, mavenFileFilter, sourceDirectory, outputDirectory);
+        MojoParameters params = createMojoParameters();
         File dockerArchive = dockerAssemblyManager.createDockerTarArchive(imageName,params, imageConfig.getBuildConfiguration());
 
         dockerAccess.buildImage(imageName, dockerArchive);
@@ -97,4 +101,8 @@ public class BuildMojo extends AbstractDockerMojo {
         }
     }
 
+    private MojoParameters createMojoParameters() {
+        return new MojoParameters(session, project, archive, mavenFileFilter, mavenFilterReader,
+                sourceDirectory, outputDirectory);
+    }
 }