+++ src/main/java/org/jolokia/docker/maven/assembly/DockerAssemblyConfigurationSource.java
@@ -229,7 +229,7 @@ public class DockerAssemblyConfigurationSource implements AssemblerConfiguration
 
     @Override
     public MavenReaderFilter getMavenReaderFilter() {
-        return null;
+        return params.getMavenFilterReader();
     }
 
     @Override