+++ doc/changelog.md
@@ -2,6 +2,7 @@
 
 * **0.11.2**
   - Fix maven parse error when specifying restart policy (#99)
+  - Allow hostnames to be used in port bindings
 
 * **0.11.1**
   - Add support for binding UDP ports (#83)